general_formコンポーネントを作成し、new_member_generalを変更
